requirements.txt
Flask==3.1.0
Flask_SocketIO==5.5.1
python-dotenv==1.0.1
simplejson==3.20.1
websocket_client==1.8.0
websockets==15.0.1
Pillow==11.1.0
sdnotify==0.3.2
qrcode==8.0
requests==2.32.3